+++ tests/dummy/app/controllers/examples.js
@@ -7,6 +7,7 @@ var ExamplesController = Ember.Controller.extend({
   favouritePizzaIds: null,
   preselectedPizzaIds: ["ham", "pep", "tex"],
   chosenIngredients: null,
+  enabled: true,
 
   pizzas: Ember.A([
     {
@@ -94,6 +95,10 @@ var ExamplesController = Ember.Controller.extend({
     loadMorePizzas: function() {
       this.pizzas.pushObjects(this.evenMorePizzas);
       this.evenMorePizzas = [];
+    },
+
+    toggleEnabled: function() {
+      this.toggleProperty('enabled');
     }
   }
 });