+++ tests/unit/components/select-2-test.js
@@ -62,18 +62,23 @@ var categorizedIngredients = [
 var ingredients = [
   {
     id: 1,
-    name: 'Tomato'
+    name: 'Tomato',
+    subtext: 'first'
   }, {
     id: 2,
-    name: 'Peperoni'
+    name: 'Peperoni',
+    subtext: 'second'
   }, {
     id: 3,
-    name: 'Ham'
+    name: 'Ham',
+    subtext: 'third'
   }, {
     id: 4,
-    name: 'Chorizo'
+    name: 'Chorizo',
+    subtext: 'fourth'
   }
 ];
+
 
 var App, component;
 moduleForComponent('select-2', 'Select2Component', {
@@ -553,12 +558,32 @@ test("it uses optionLabelPath", function() {
   andThen(function() {
     strictEqual(component.get('value'), ingredients[1], "selects correct item");
     equal($('.select2-chosen').text(), ingredients[1].name, "has correct text");
-
-    click('.select2-choice');
-    click('.select2-results li:nth-child(1)', 'body');
   });
 });
 
+
+test("it uses optionDescriptionPath", function() {
+  expect(1);
+
+  var component = this.subject({});
+
+  component.set('optionLabelPath', 'name');
+  component.set('optionDescriptionPath', 'subtext');
+  component.set('content', ingredients);
+
+  this.append();
+
+  // open options by clicking on the element
+  click('.select2-choice');
+
+  andThen(function() {
+    var expected = ingredients.map(function(ingredient) {
+      // jQuery .text() will have space between name and subtext, but thats ok
+      return ingredient.name + ' ' + ingredient.subtext;
+    }).join('');
+    equal($('.select2-results li').text(), expected, "display correct text");
+  });
+});
 
 test("it is disabled when `enabled=false`", function() {
   expect(3);